pyproject.toml
[project]
name = "powermeterread"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "pyodbc (>=5.2.0,<6.0.0)"
]

[tool.poetry]
packages = [{include = "powermeterread", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
